chore: some minor tweaks
