+++ packages/oas-utils/src/entities/spec/spec-objects.ts
@@ -1,4 +1,4 @@
-import { type ENTITY_BRANDS, nanoidSchema } from '@scalar/types/entities'
+import { type ENTITY_BRANDS, nanoidSchema } from '@scalar/types/utils'
 import { z } from 'zod'
 
 /**
