+++ packages/types/src/entities/index.ts
@@ -1,4 +1,3 @@
-export { type ENTITY_BRANDS, nanoidSchema } from './utility.ts'
 export {
   type Oauth2Flow,
   type Oauth2FlowPayload,
