+++ packages/types/tsconfig.json
@@ -2,6 +2,7 @@
   "extends": "../../tsconfig.node.json",
   "include": ["src", "test"],
   "compilerOptions": {
+    "rootDir": "./src",
     "paths": {
       "@/*": ["./src/*"],
       "@test/*": ["./test/*"]
